requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
langchain-community
pypdf
bs4
arxiv
pymupdf
fitz
wikipedia
langchain-text-splitters
langchain_openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain_ollama
langchain_groq
fastapi
pip install langserve
